backend/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
validators==0.20.0
pymongo==4.1.1
python-dotenv==0.19.0
dnspython==2.2.1
flask_cors